apps/api/requirements.txt
# Minimal deps for the current stub (no LLM/RAG yet). Use a venv to avoid conflicts with other projects.
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.8.0
python-dotenv>=1.0.0
httpx>=0.25.0
sse-starlette>=1.8.0

# RAG retrieval (Postgres + pgvector + OpenAI embeddings)
psycopg[binary]>=3.1.0
pgvector>=0.2.0
openai>=1.12.0
langgraph>=0.2.0

# Optional: live PartSelect fetch when cache/DB miss (search_parts by model_number)
playwright>=1.40.0
